+++ NEWS.md
@@ -1,13 +1,16 @@
 # metaUI 0.1.2 (under development)
-* Added the ability to create an app without any filters/moderators
 
-## Minor changes
-* generate_shiny() has gained an options argument to allow for more customization of the app. Currently it allows setting a Shiny theme, and a limit for the number of rows in a forest plot, with sensible defaults.
+## Minor enhancements
+* generate_shiny() has gained an `options` argument to allow for more customization of the app. Currently it allows setting a Shiny theme, and a limit for the number of rows in a forest plot, with sensible defaults.
+* prepare_data gained an argument `arrange_filters` to allow for the ordering of filters in the app. It now defaults to ordering them in the order they are passed into the function, but this can be changed to alphabetical or no ordering (in which case the order of columns determines their position)
+* Added the ability to create an app without any filters/moderators
+* Now shows "(Missing)" level only where there are any missing values in that filter/moderator - unless
 
 ## Bug fixes
 * Corrected random intercept specification in rma.mv (and added sparse = TRUE to speed up model fitting)
 * Fixed bug in describing moderators where "Other"-category already existed
 * Fixed creation of code to install required packages and added check to ensure that filters are factors or numeric (#28)
+* Waffle plots in sample description no longer run out of colors in the presence of 11 categories.
 
 # metaUI 0.1.1